``` refactor(api): rename persistence sdk package to sdk
The package name org.papertrail.persistencesdk was renamed
to org.papertrail.sdk. This change simplifies the SDK's
identifier, making it more concise and general.
It prepares the SDK for potential future expansions beyond
persistence-specific functionalities, enhancing its clarity
and consistency within the project.
```
