+++ src/main/java/org/papertrail/sdk/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Embedded SDK for resolving calls to the PaperTrail PersistenceAPI
  */
-package org.papertrail.persistencesdk;
+package org.papertrail.sdk;
